chore: move NoServiceError and lint
